+++ src/components/volunteers/Volunteer/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { VolunteerResponse } from '@/types/dataModel/volunteer';
 import { Typography, Box } from '@mui/material';
 import { VolunteerEventResponse } from '@/types/dataModel/eventVolunteer';
-import { getRoleIcons } from '@/utils/role';
+import IconList from '@/components/IconList';
 
 // Use VolunteerResponse Props
 interface VolunteerProps {
@@ -92,7 +92,7 @@ export default function Volunteer({
           </Typography>
           {events.map((volunteerEvent, i) => (
             <Box key={i} sx={{ display: 'flex' }} pt={1}>
-              {getRoleIcons([volunteerEvent.role])}
+              <IconList roles={[volunteerEvent.role]}></IconList>
               <Typography pl={2}>{volunteerEvent.event.name}</Typography>
             </Box>
           ))}
